project_echo/pyproject.toml
[tool.poetry]
name = "project_echo"
version = "0.1.0"
description = "Hierarchical Image Memory reference implementation"
authors = ["AutoGen Agent <agent@example.com>"]
packages = [
    {include = "him"},
    {include = "sel_bot"},
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {version = "^0.27.0", extras = ["standard"]}
pydantic = "^2.6.3"
blake3 = "^1.0.7"
"discord.py" = "^2.3.2"
httpx = "^0.27.0"
sqlalchemy = "^2.0.28"
aiosqlite = "^0.19.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
asyncpg = ">=0.29,<0.32"
langchain = "^0.3.0"
pillow = "^11.0.0"
numpy = "^1.24.0"
pynacl = "^1.6.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.25.0"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
